Rename directories and fargate Go module
