+++ rehydrate/README.md
