+++ rehydrate/fargate/Dockerfile
