+++ rehydrate/fargate/objectProcessor.go
